Fixes #9 - find definition
This is an experimental implementation of a function to find where
a keyword is defined. Highlight a keyword name and select "Find
Definition" from the robot menu and it will run the brackets "Find
in files" command with a regexp that should match only where a
keyword is defined.

I may replace this with a better implementation, but it's surprisingly
useful as-is.